chore: make constant shared
