+++ libs/application/templates/inao/financial-statement-cemetery/src/fields/PowerOfAttorney/index.tsx
@@ -17,7 +17,7 @@ import { m } from '../../lib/messages'
 import { FieldBaseProps } from '@island.is/application/types'
 import { getErrorViaPath } from '@island.is/application/core'
 import { IdentityQuery } from '../../graphql'
-import { ABOUTIDS } from '../../utils/constants'
+import { ABOUTIDS } from '../../../../shared/utils/constants'
 
 export const PowerOfAttorneyFields = ({ application }: FieldBaseProps) => {
   const { formatMessage } = useLocale()
