+++ libs/application/templates/inao/financial-statement-individual-election/src/fields/ElectionInfoFields/index.tsx
@@ -4,17 +4,15 @@ import { useLocale } from '@island.is/localization'
 
 import { useQuery } from '@apollo/client'
 import { getAvailableElections } from '../../graphql'
-import {
-  Application,
-  FinancialStatementsInaoElection,
-} from '@island.is/api/schema'
+import { FinancialStatementsInaoElection } from '@island.is/api/schema'
 import { useFormContext } from 'react-hook-form'
-import { ABOUTIDS, UPDATE_ELECTION_ACTION } from '../../lib/utils/constants'
+import { UPDATE_ELECTION_ACTION } from '../../lib/utils/constants'
 import { electionInitialState, electionReducer } from './electionReducer'
 import { Options } from '../../types/types'
 import { ElectionSelect } from './ElectionSelect'
 import { IncomeLimitFields } from './IncomeLimitFields'
 import { m } from '../../lib/utils/messages'
+import { ABOUTIDS } from '../../../../shared/utils/constants'
 
 export const ElectionsInfoFields = () => {
   const [state, dispatch] = useReducer(electionReducer, electionInitialState)
