+++ libs/application/templates/inao/financial-statement-individual-election/src/fields/Overview/index.tsx
@@ -18,7 +18,7 @@ import { formatCurrency } from '../../../../shared/utils/formatCurrency'
 import { starterColumnStyle, sectionColumn } from './overviewStyles.css'
 import { FinancialStatementIndividualElection } from '../../lib/utils/dataSchema'
 import { useSubmitApplication } from '../../hooks/useSubmitApplication'
-import { GREATER } from '../../lib/utils/constants'
+import { GREATER } from '../../../../shared/utils/constants'
 import { m } from '../../lib/utils/messages'
 import { CapitalNumberOverview } from '../../../../shared/components/CapitalNumberOverview'
 import { AssetDebtEquityOverview } from '../../../../shared/components/AssetDebtEquityOverview'
