+++ libs/application/templates/inao/financial-statement-individual-election/src/forms/applicationForm/individualKeyNumbersSection/capitalNumbersSection.ts
@@ -4,7 +4,7 @@ import {
   buildSubSection,
 } from '@island.is/application/core'
 import { m } from '../../../lib/utils/messages'
-import { CAPITALNUMBERS } from '../../../lib/utils/constants'
+import { CAPITALNUMBERS } from '../../../../../shared/utils/constants'
 
 export const capitalNumberSection = buildSubSection({
   id: 'keynumbers.capitalNumbers',
