+++ libs/application/templates/inao/financial-statement-individual-election/src/forms/applicationForm/overviewSection/index.ts
@@ -4,7 +4,7 @@ import {
   buildSection,
   getValueViaPath,
 } from '@island.is/application/core'
-import { ABOUTIDS, GREATER, LESS } from '../../../lib/utils/constants'
+import { ABOUTIDS, GREATER, LESS } from '../../../../../shared/utils/constants'
 import { m } from '../../../lib/utils/messages'
 
 export const overviewSection = buildSection({
