+++ libs/application/templates/inao/financial-statement-individual-election/src/forms/done/conclusionSection.ts
@@ -4,9 +4,6 @@ import {
   buildSection,
 } from '@island.is/application/core'
 import { m } from '../../lib/utils/messages'
-import { FSIUSERTYPE } from '../../types/types'
-import { Application } from '@island.is/application/types'
-import { getCurrentUserType } from '../../lib/utils/helpers'
 
 export const conclusionSection = buildSection({
   id: 'conclusionSection',
@@ -14,12 +11,7 @@ export const conclusionSection = buildSection({
   children: [
     buildMultiField({
       id: 'conclusion',
-      title: (application: Application) => {
-        const answers = application.answers
-        const externalData = application.externalData
-        const userType = getCurrentUserType(answers, externalData)
-        return userType === FSIUSERTYPE.INDIVIDUAL ? m.infoReceived : m.received
-      },
+      title: m.infoReceived,
       children: [
         buildCustomField({
           id: 'overview',
