+++ libs/application/templates/inao/financial-statement-individual-election/src/lib/utils/constants.ts
@@ -1,21 +1,4 @@
-export const ABOUTIDS = {
-  operatingYear: 'conditionalAbout.operatingYear',
-  applicationType: 'conditionalAbout.applicationType',
-  selectElection: 'election.selectElection',
-  electionName: 'election.electionName',
-  electionId: 'election.electionId',
-  genitiveName: 'election.genitiveName',
-  incomeLimit: 'election.incomeLimit',
-  powerOfAttorneyNationalId: 'about.powerOfAttorneyNationalId',
-  powerOfAttorneyName: 'about.powerOfAttorneyName',
-}
-
-export const GREATER = 'greater'
-export const LESS = 'less'
-export const VALIDATOR = 'validator'
 export const UPDATE_ELECTION_ACTION = 'updateCurrentElection'
-export const INPUTCHANGEINTERVAL = 300
-export const TOTAL = 'total'
 export const ELECTIONLIMIT = 550000
 
 export const INDIVIDUALOPERATIONIDS = {
@@ -33,32 +16,4 @@ export const INDIVIDUALOPERATIONIDS = {
   capitalCost: 'individualExpense.capitalCost',
   totalIncome: 'individualIncome.total',
   totalExpense: 'individualExpense.total',
-}
-
-export const EQUITIESANDLIABILITIESIDS = {
-  assetPrefix: 'asset',
-  currentAssets: 'asset.currentAssets',
-  fixedAssetsTotal: 'asset.fixedAssetsTotal',
-  assetTotal: 'asset.total',
-  liabilityPrefix: 'liability',
-  longTerm: 'liability.longTerm',
-  shortTerm: 'liability.shortTerm',
-  asset: 'liability.asset',
-  totalLiability: 'liability.total',
-  operationResult: 'equity.operationResult',
-  equityPrefix: 'equity',
-  totalEquity: 'equity.totalEquity',
-  totalCash: 'equity.total',
-  totalEquityAndLiabilities: 'equityAndLiabilities.total',
-}
-
-export const CAPITALNUMBERS = {
-  capitalPrefix: 'capitalNumbers',
-  capitalIncome: 'capitalNumbers.capitalIncome',
-  capitalCost: 'capitalNumbers.capitalCost',
-  total: 'capitalNumbers.total',
-}
-
-export const OPERATINGCOST = {
-  total: 'operatingCost.total',
 }
