+++ libs/application/templates/inao/financial-statement-political-party/src/fields/PowerOfAttorneyFields/PowerOfAttorneyFields.tsx
@@ -13,7 +13,7 @@ import { useLocale } from '@island.is/localization'
 import { IdentityInput, Query } from '@island.is/api/schema'
 
 import { m } from '../../lib/messages'
-import { ABOUTIDS } from '../../utils/constants'
+import { ABOUTIDS } from '../../../../shared/utils/constants'
 import { FieldBaseProps } from '@island.is/application/types'
 import { getErrorViaPath } from '@island.is/application/core'
 import { IdentityQuery } from '../../graphql'
