+++ libs/application/templates/inao/financial-statement-political-party/src/forms/applicationForm/clientInfoSection/index.ts
@@ -6,7 +6,7 @@ import {
   buildTextField,
 } from '@island.is/application/core'
 import { m } from '../../../lib/messages'
-import { ABOUTIDS } from '../../../utils/constants'
+import { ABOUTIDS } from '../../../../../shared/utils/constants'
 import { Application, UserProfile } from '@island.is/application/types'
 import { Identity } from '@island.is/clients/identity'
