+++ libs/application/templates/inao/financial-statement-political-party/src/forms/applicationForm/keyNumbersSection/financialStatementSection/index.ts
@@ -4,7 +4,7 @@ import {
   getValueViaPath,
 } from '@island.is/application/core'
 import { m } from '../../../../lib/messages'
-import { LESS } from '../../../../utils/constants'
+import { LESS } from '../../../../../../shared/utils/constants'
 
 export const financialStatementSection = buildSection({
   id: 'documents',
