+++ libs/application/templates/inao/financial-statement-political-party/src/utils/constants.ts
@@ -1,21 +1,6 @@
 export const PartiesYearAllowed = 'PartiesYearAllowed'
 export const PartiesBackwardLimit = 'PartiesBackwardLimit'
-export const INPUTCHANGEINTERVAL = 300
-export const TOTAL = 'total'
-export const LESS = 'less'
-export const GREATER = 'greater'
 export const ELECTIONLIMIT = 550000
-
-export const ABOUTIDS = {
-  operatingYear: 'conditionalAbout.operatingYear',
-  applicationType: 'conditionalAbout.applicationType',
-  selectElection: 'election.selectElection',
-  electionName: 'election.electionName',
-  genitiveName: 'election.genitiveName',
-  incomeLimit: 'election.incomeLimit',
-  powerOfAttorneyNationalId: 'about.powerOfAttorneyNationalId',
-  powerOfAttorneyName: 'about.powerOfAttorneyName',
-}
 
 export const PARTYOPERATIONIDS = {
   incomePrefix: 'partyIncome',
@@ -34,34 +19,3 @@ export const PARTYOPERATIONIDS = {
   capitalCost: 'partyExpense.capitalCost',
   totalExpense: 'partyExpense.total',
 }
-
-export const OPERATINGCOST = {
-  total: 'operatingCost.total',
-}
-
-export const CAPITALNUMBERS = {
-  capitalPrefix: 'capitalNumbers',
-  capitalIncome: 'capitalNumbers.capitalIncome',
-  capitalCost: 'capitalNumbers.capitalCost',
-  total: 'capitalNumbers.total',
-}
-
-export const EQUITIESANDLIABILITIESIDS = {
-  assetPrefix: 'asset',
-  currentAssets: 'asset.currentAssets',
-  fixedAssetsTotal: 'asset.fixedAssetsTotal',
-  assetTotal: 'asset.total',
-  liabilityPrefix: 'liability',
-  longTerm: 'liability.longTerm',
-  shortTerm: 'liability.shortTerm',
-  asset: 'liability.asset',
-  totalLiability: 'liability.total',
-  operationResult: 'equity.operationResult',
-  equityPrefix: 'equity',
-  totalEquity: 'equity.totalEquity',
-  totalCash: 'equity.total',
-  totalEquityAndLiabilities: 'equityAndLiabilities.total',
-}
-
-// Error helpers
-export const VALIDATOR = 'validator'
