+++ libs/application/templates/inao/shared/utils/constants.ts
@@ -1,0 +1,50 @@
+export const INPUTCHANGEINTERVAL = 300
+export const TOTAL = 'total'
+export const GREATER = 'greater'
+export const LESS = 'less'
+
+export const CAPITALNUMBERS = {
+  capitalPrefix: 'capitalNumbers',
+  capitalIncome: 'capitalNumbers.capitalIncome',
+  capitalCost: 'capitalNumbers.capitalCost',
+  total: 'capitalNumbers.total',
+}
+
+export const ABOUTIDS = {
+  operatingYear: 'conditionalAbout.operatingYear',
+  applicationType: 'conditionalAbout.applicationType',
+  selectElection: 'election.selectElection',
+  electionName: 'election.electionName',
+  electionId: 'election.electionId',
+  genitiveName: 'election.genitiveName',
+  incomeLimit: 'election.incomeLimit',
+  powerOfAttorneyNationalId: 'about.powerOfAttorneyNationalId',
+  powerOfAttorneyName: 'about.powerOfAttorneyName',
+}
+
+export const EQUITIESANDLIABILITIESIDS = {
+  assetPrefix: 'asset',
+  currentAssets: 'asset.currentAssets',
+  fixedAssetsTotal: 'asset.fixedAssetsTotal',
+  assetTotal: 'asset.total',
+  liabilityPrefix: 'liability',
+  longTerm: 'liability.longTerm',
+  shortTerm: 'liability.shortTerm',
+  asset: 'liability.asset',
+  totalLiability: 'liability.total',
+  operationResult: 'equity.operationResult',
+  equityPrefix: 'equity',
+  totalEquity: 'equity.totalEquity',
+  totalCash: 'equity.total',
+  totalEquityAndLiabilities: 'equityAndLiabilities.total',
+}
+
+export const OPERATINGCOST = {
+  total: 'operatingCost.total',
+}
+
+// Error helpers
+export const VALIDATOR = 'validator'
+export const APPLICANTASMEMBER = 'applicantasmember'
+export const ACTORASCARETAKER = 'actorascaretaker'
+export const ACTORLONEBOARDMEMBER = 'actorloneboardmember'
